Fix executing without auth in tests
Because Client.auth_required is a property, when mocking it
mock.PropertyMock must be used.
This bug was invisible because of mocking requests library
in all places where fixed mock was used and call to verison API
was also mocked. This call was returning malformed body so it was
treated as auth is not required.

Change-Id: I634e56bbbc38f59dbea29e39536403480c90f013