Add possibility of editing cluster's attributes
Since Fuel 6.1 we have ability to add external repositories and change
them via UI. The set of repos are the part of cluster's attributes, so
in order to change them we have to provide the ability of changing any
of them.

This patch adds ability to download and upload cluster attributes
into/from yaml file, just like it's implemented for deployment tasks.

Implements: blueprint consume-external-ubuntu
Change-Id: I7a17d8da299b43940f20572d44615c21036538ab