Use compatible version of six
Fuel Client uses features of six that appeared
in the version 1.7 but requires a version which
is >= 1.5.2. That causes it to fail if it's installed
to an environment with an old version of six.

This patch changes the version specification to
six>=1.7.0,<=1.9.0 which is compatible with the
Juno release of OpenStack.

Partial-bug: #1439089
Change-Id: Id06ba598df6babe0b5ffcf0a7b08049dbbeaf429