Return realistic data from mocked library
Base CLI test mocks client library in order to test
command line components indepentently. If not configured
explicitly a mocked library returns mocks that accept
any incomng message without producing an error, even if
in real life scenarions if should be risen.

This patch configured mocked library to return realistic
data and fixes a problem in the code that was missed
because of the above-mentioned misconfiguration.

Closes-bug: #1467510
Change-Id: I065345d59c94e06f9d061c2e9b0b415e45fd62fb