py3: Fix tox -e py34
Run a subset of tests on Python 3.4: only run test_utils currently.
With a working "tox -e py34", the py34 check job can be made voting.
A voting check job ensures that no code incompatible with Python 3
will be added back to the code tested by "tox -e py34". It avoids the
risk of having to port the same code twice to Python 3.

When more tests are ported to Python 3.4, the whitelist of tests
will grow slowly. The goal is to remove the [testenv:py34] section
from tox.ini to run all tests on Python 3.4.

Change-Id: I164d2ff3d948c8339edcfd96abc46d869018a324