Error message when settings file doesn't exist
Catch possible exceptions which relate to a situation when the
directory (provided by a user) doesn't exist or the directory doesn't
have sufficient permissions or when it is not directory at all. Catch
these exceptions and display appropriate error messages. When debug mode
is turned on then display stack traceback.

Closes-bug: #1330429
Change-Id: I3551dc42e527d60a2d8b2040418b2594e08e3b09