Generate user settings file
Since relying on a system-wide /etc/... configuration
file is not a good approach for a user-side application
it's necessary to not use that file by default.

This patch changes default settings file to ~/.config/fuel_client.yaml
and adds generation of one with default values if it does not exist.
Specifying of a custom settings file using FUELCLIENT_CUSTOM_SETTINGS
variable is still possible.

DocImpact
Closes-bug: #1458361
Change-Id: Ia17c8edbdd71a74318d3237a47d4b5650934c378