Add check if task is in error before downloading snapshot
Checking of status of generate snapshot task added before actual
download action in get_snapshot method. This eliminates the issue
with ambiguos error from download_snapshot_with_progress_bar function
that occurs because task which ended with error contains unexpected by
the code data to process (i.e. 'message' field is used to form url for
http request but in the case it contains additional description of the
error).

Change-Id: I4a4754916c43efc2acb06f40cf2dc1df6bb4e9f5
Closes-Bug: #1540441