Update sync_deploymnet_tasks messages
The help message for sync-deployment-tasks does a poor job showing
the relation between the release string and the necessary directory
structure to function.

At the same time, the message for when tasks where not found for a
release implies that there is a problem that must be resolved.

Here we update the strings to hopfully be more clear about what
is going on.

Change-Id: Ib5465709aaed7ac461b75f60a0849fbae11d482d
Closes-bug: #1546321