Remove legacy patching code
In Fuel 5.1 we had an experimental feature - 'patching openstack env'.
The idea was to update and to rollback OpenStack environments between
minor releases. However, we have encounter a lot of problems with
restoring OpenStack databases andresolving dependency hell in packages,
so we buried it and never release it.

This patch removes legacy code from the source tree. We can do it
without fear, since it wasn't released in public and, moreover,
was broken since Fuel 6.0 (no chance someone use it).

Change-Id: I38625a354ea1f09efa24f3e4d18cd4337b7dcc5b
Related-Bug: #1511499