Use FileFormatBasedSerializer for 'upload' method of VIPAction
Base fuel client serializer (which is used for v1 of codebase) while
reading given file assumes that it is supplied w/o file format
extension and adds it to passed string before actual read operation.
This leads to obvious error in case of VIP related actions as
full path to a file is used as argument for those. Using
FileFormatBasedSerializer eliminates such issue as its 'read_from_file'
methods operates on given path as it is.

Change-Id: I0a954f6c5c91130a0e3cd5e0dfa896b2adc2d6d3
Closes-Bug: #1553241