Check response from version for auth required
If the version call requires auth, we need to catch that to understand
if we should use auth for making requests.  Previously the version api
call may have been open which allowed the CLI to make the call to
determine if auth should be used. If the version call itself requires
auth, then we need to check the response code to enable authentication
for the api requests.

Change-Id: Ie70d601eb4db4909e52e6afaa15cba6195f24045
Related-Bug: #1585137