Fix fuel cli output function
FUEL CLI output function produces GET requests to the nodes belongs to the task to fetch
their progress. But, node is changing its id during the node's removing task and we may
encounter case when we are requesting the progress for already deleted node. This patch
introduces check for node existence.

Change-Id: Ia85db174976cb0dbacaf20e6378fcbb9f1ed7cf0
Closes-Bug: #1565026