Fix problems with os-username and os-password usage
In case, when we import DefaultAPIClient, it tries to instantiate
APIClient.default_client with settings object, but at time of initialization
username and password parameters are not being parsed yet. It causes situation,
that it always instantiates with default credentials.
In order to prevent it, lazy initializer was created. APIClient.default_client
will be executed on first actual usage which is performed after parameters
parsing.

Closes-Bug: #1593721
Change-Id: Ic0608f11a5db60691f766608fe8eee578c9fb0c6