Add more verbosity to fuel2 vms commands
There is a set of commands that operate on vms.
But the API and respective fuel2 commands are
only available if 'advanced' feature is active
in 'feature_group' (from fuelmenu). In other case
Nailgun reports '(not found)'. This patch adds more
verbosity to fuel2 vms commands if 'advanced' feature
is not active instead '(not found)' responce.

Change-Id: Idc1a2612ca20125ed54bdd35bfb048a81fa22cdd
Closes-Bug: 1603980